statement view
