--- components/workflow-view/runs-info/index.ts
@@ -1,2 +1,0 @@
-import { RunsInfo } from "./runs-info";
-export default RunsInfo;
