--- components/workflow-view/runs-info/list-items/index.ts
@@ -1,2 +1,0 @@
-import { ListItems } from "./list-items";
-export default ListItems;
