--- components/workflow-view/runs-info/runInfo/index.ts
@@ -1,2 +1,0 @@
-import { RunInfo } from "./runInfo";
-export default RunInfo;
