+++ components/workflow-view/summary/summary.tsx
@@ -1,5 +1,5 @@
 import { Typography, Card, CardContent, Grid } from "@mui/material";
-import WorkflowRunStatus from "../runs-info/run-status";
+import WorkflowRunStatus from "../workflow-runs/detail/run-status";
 import { WorkflowItem } from "../../../gen/controlplane/v1/response_messages";
 
 export const WorkflowSummary = ({ wf }: { wf: WorkflowItem }) => {
