+++ components/workflow-view/workflow-runs/detail/attestation-info/index.ts
@@ -1,0 +1,2 @@
+import { AttestationInfo } from "./attestation-info";
+export default AttestationInfo
