+++ components/workflow-view/workflow-runs/detail/index.ts
@@ -1,0 +1,2 @@
+import { RunDetail } from "./detail";
+export default RunDetail;
