+++ components/workflow-view/workflow-runs/detail/summary/index.ts
@@ -1,0 +1,2 @@
+import { WorkflowRunSummary } from "./summary";
+export default WorkflowRunSummary
