+++ components/workflow-view/workflow-runs/index.ts
@@ -1,0 +1,2 @@
+import { WorkflowRuns } from "./workflow-runs";
+export default WorkflowRuns;
