+++ components/workflow-view/workflow-runs/list-sidebar/index.ts
@@ -1,0 +1,2 @@
+import { RunsListSidebar } from "./list-sidebar";
+export default RunsListSidebar;
