+++ components/workflow-view/workflow-runs/list-sidebar/list-sidebar.tsx
@@ -13,7 +13,7 @@ import AutorenewIcon from "@mui/icons-material/Autorenew";
 import { ReactNode, useEffect, useState } from "react";
 import { WorkflowRunItem } from "../../../../gen/controlplane/v1/response_messages";
 
-export const ListItems = ({
+export const RunsListSidebar = ({
   runs,
   onSelect,
 }: {
