+++ components/workflow-view/workflow-runs/workflow-runs.tsx
@@ -7,14 +7,14 @@ import {
   Typography,
 } from "@mui/material";
 import { useState } from "react";
-import ListItems from "./list-items";
-import RunInfo from "./runInfo";
+import RunsListSidebar from "./list-sidebar";
+import RunDetail from "./detail";
 import { useAuth } from "../../../contexts/auth";
 import { useWorkflowRunsList } from "../../../lib/apiclient/workflowRuns";
 import WithLoader from "../../with-loader";
 import { WorkflowRunItem } from "../../../gen/controlplane/v1/response_messages";
 
-export const RunsInfo = ({ workflowID }: { workflowID: string }) => {
+export const WorkflowRuns = ({ workflowID }: { workflowID: string }) => {
   const [currentRunID, setCurrentRunID] = useState("");
   const { apiClient } = useAuth();
   const { isLoading, data } = useWorkflowRunsList(workflowID, apiClient);
@@ -28,10 +28,13 @@ export const RunsInfo = ({ workflowID }: { workflowID: string }) => {
           {data?.result && data.result.length > 0 && (
             <Grid container>
               <Grid item xs={3} xl={2}>
-                <ListItems runs={data?.result!} onSelect={setCurrentRunID} />
+                <RunsListSidebar
+                  runs={data?.result!}
+                  onSelect={setCurrentRunID}
+                />
               </Grid>
               <Grid item xs={9} xl={10}>
-                <RunInfo runID={currentRunID} />
+                <RunDetail runID={currentRunID} />
               </Grid>
             </Grid>
           )}
